+++ src/Ydb.Sdk/src/Services/Topic/Reader/Utils.cs
@@ -6,4 +6,10 @@ internal static class Utils
     {
         return bytesSize / countParts + (currentIndex == countParts - 1 ? bytesSize % countParts : 0);
     }
+    
+    internal static void SetPartitionClosedException(CommitSending commitSending, long partitionSessionId)
+    {
+        commitSending.TcsCommit.TrySetException(
+            new ReaderException($"PartitionSession[{partitionSessionId}] was closed by server."));
+    }
 }