+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed TopicSession.ReconnectSession(): CompareExchange returns the original value in location1.
 * Fixed: YdbDataReader.GetDataTypeName for optional values.
 * Added support for "Columns" collectionName in YdbConnection.GetSchema(Async).